backend/app/pyproject.toml
[project]
name = "vallyChat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt==4.3.0",
    "bidict==0.23.1",
    "bleach>=6.3.0",
    "blinker==1.9.0",
    "click==8.3.1",
    "cuid2==2.0.1",
    "dnspython==2.8.0",
    "eventlet==0.40.4",
    "flask==3.1.2",
    "flask-cors==6.0.2",
    "flask-socketio==5.6.0",
    "gevent==25.9.1",
    "gevent-websocket==0.10.1",
    "greenlet==3.3.0",
    "gunicorn==23.0.0",
    "h11==0.16.0",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "loguru==0.7.3",
    "markdown>=3.10.2",
    "markupsafe==3.0.3",
    "packaging==25.0",
    "pyjwt==2.8.0",
    "pymysql==1.1.2",
    "python-dotenv==1.0.1",
    "python-engineio==4.12.3",
    "python-socketio==5.15.1",
    "redis==7.1.0",
    "simple-websocket==1.1.0",
    "sqlalchemy==2.0.44",
    "typing-extensions==4.15.0",
    "werkzeug==3.1.5",
    "wsproto==1.3.2",
    "zope-event==6.1",
    "zope-interface==8.2",
]
